apps/devops/FlaskBack4AppDocker/requirements.txt
requests
Flask
werkzeug
sendgrid
faker
pytz
sqlalchemy
pyodbc
newsapi-python
eventbrite
scrapetube
pymysql
schedule
sentry-sdk
Gunicorn


pytest
coverage
requests-mock
pytest-watch
pytest-mock
boto3
